clarify 2024.1 testing runtime requirements
ubuntu 20.04 was the default testing runtime used
for the Zed release. It was then included as
additional testing for a smooth upgrade in 2023.1.

ubuntu 20.04 was removed from the testing runtime in 2023.2
and with it, the last use of Python 3.8 was also removed.
This resulted in some projects dropping support which broke
others that still tested on Python 3.8. As a result of this
I8ecf4acc658854b971c3d6b855bb84d6f5932f85 reintroduced py3.8
as a testing requirement for 2023.2.
Additionally, in the 2024.1 runtimes document a clarifying
statement was added stating Python projects should avoid
removing support for non-EOL Python version referencing
a change in the PTI requirements introduced by
Ie2bb6c96e39eb21ab42132774699f254c7f476a9
This change amends the 2024.1 runtimes to make it clear
that while py38 testing for language support is expected
there is no expectation that Ubuntu 20.04 will be tested or
supported. Ubuntu 20.04 can be used to fulfil the python
3.8 requirements for tox jobs and it is only listed
because it is what we are currently using for this
requirement.

As we are moving forward to supporting Python 3.10, 3.11
and 3.12 upstream Python is actively removing features that
were deprecated in 3.8. The upstream Python community realised
that the larger projects could not keep up with
the rate of change and have adopted a longer deprecation period
with faster release. The new Python release cycle is now yearly,
with a predictable release in October of each year. The deprecation period is now 2 Python releases and removal in the third release.

This nicely aligns with the SLURP cadence allowing OpenStack
to declare the deprecation of a min python version in a slurp
release and the removal in the following non-slurp.

For 2024.1 this patch declares that 2024.1 will be the last
release to support Python 3.8 in advance of its EOL in
October 2024. It also declares Python 3.9 as the next
min python version for 2024.2 based on its usage in Centos 9
stream, Debian 11 and the upstream support EOL date of October
2025.

Assuming the same cadence continues this will allow OpenStack to
Deprecate 3.9 support in the 2025.1 runtime and remove
it in the 2025.2 runtime. and continue raising our minimum Python version
on a predictable yearly cadence in lockstep with the upstream Python
release schedule.

Change-Id: Ifa2479de24414750f7fac421489540e73a56ed7c