[docs] Add links to release based runtimes
We have the release based runtimes documented in this repo and
published.  However, The link to them from within the PTI page is
easily missed.  This change adds a specific runtimes index page and
links to that from the reference section of the governance homepage.

The list included in the PTI page uses the default ascii sorting i.e.
2023.1, 2023.2, 2024.1, Stein, ..., Zed

The change also updates this list to follow the release order.

Change-Id: I0b8cd23986f6abb168486f8cf2a8c30e2a1b30e9