Resolution to create openstack-unmaintained core
This patch proposes an amendment to the 2023-07-24 "Unmaintained status
replaces Extended Maintenance" resolution to allow an openstack-wide
group to handle Unmaintained branch changes across all projects by
default, though this can be overridden by individual projects.

Additionally, the original resolution file has been modified to
include a note that the amendment exists along with a link to it.

Change-Id: Iecc9751091285a5d3ca72c8d959136b33e274839