Move note closer to amended text
Change Iecc9751091285a5d3c amended resolution 2023-07-24, but left
a note in the original resolution that is far enough away from the
amended point to be confusing.  So move it to a better location.

Also put the reference list into a line block so it displays
as intended.

These are formatting changes only; there's no change to the text.

Change-Id: Ide5068e20725cf46e3204c0fc3b2c84d99c8ba80