Remove Eventlet From Openstack
This is a new community goal proposal that aim to propose
solutions to address the current eventlet issues discussed
in the mailing list [1].

The goal of this proposal is to find a solution to remove our
dependence on Eventlet, and so to remove Eventlet from Openstack.

This proposal aim to propose short, medium, and long terms
solutions.

[1] https://lists.openstack.org/archives/list/openstack-discuss@lists.openstack.org/thread/YO5CZDVAJ6QSF734ALWSGNOQDDAIOXKI/#RZFUTBFTUCSHKVA6SOCWWIXEH3QJHMYT

Change-Id: I4d2c8a40c92249a32baa8e85a4251982543e2a60