OpenStack TC charter updates for the bylaws changes
There are updates to the OpenInfra Foundation bylaws that
cleans up some of the old special-case text for the OpenStack
project. Some of the changes in bylaws are important to the
OpenStack governance (for example, diversity requirement) so
we are moving those changes in the OpenStack TC charter.

OpenInfra Board has approved the bylaws changes as sent in
the below email attachment:
- https://lists.openinfra.dev/archives/list/foundation@lists.openinfra.dev/thread/6FBOOK7WVBRHYOBQPTF2T2CNJOO7XMSQ/
- https://board.openinfra.dev/meetings/2023-12-05

That attachment shows the bylaws changes and what needs to be
moved to OpenStack TC charter or to the Openinfra Project
Confirmation Guidelines.

This commit update the OpenStack TC charter for the required
changes in the bylaws.

Change-Id: Iaff919fc99b8270bc7562a61ea73c5304a0f4acf