Move yaql to Heat's governance
The yaql library has been maintained under murano's governance, but
the murano project has been inactive for several cycles. Because this
library is used by heat and mistral, we have to maintain it as long
as these two projects are maintained.

This proposes moving the yaql project to heat's governance as the first
step. Later we'll update the core reviewer of the repository so that
both heat team and mistral team can maintain it.

This transition was earlier proposed in the mailing list[1] but we
haven't received any feedback from the murano project, unfortunately.

[1] https://lists.openstack.org/archives/list/openstack-discuss@lists.openstack.org/message/KFTPG45LEB5QYY4NELA7EQJ56MSOWULK/

Change-Id: If5d3ed3a8314a8f650cd90416b53e02a84e4ed44