Retire OpenStack-Chef project
OpenStack-Chef project which provides the Chef cookbooks
for OpenStack life cycle management, does not have maintainers
and leaders to keep this project healthy. During the Caracal development
cycle election, this project is on the leaderless list.

- https://etherpad.opendev.org/p/2024.1-leaderless

A few of the stats about this project are below, which are discussed
during leadership project task:

- gate health unclear, no activity for > 1y in openstack/openstack-chef repo
- most recent branch is stable/train
- zuul config errors fixed by frickler

Previous PTL Lance Albertson indicated that they would not be able to
continue its maintenance and are ok with the retirement. (notes in above
etherpad)

Seeing no leader or maintainers for this project and its recent activities,
this change proposes to retire this project.

Change-Id: Icaf68a43028c9c49bc2bbb9dadcc9a1550bc9168