Clarify the timeline for project to move to Inactive state
Currently we try to monitor and move the inactive project to
Inactive state before milestone-2 which is deadline from the
release team. But there are chances that either TC detect the
inactive projects late or project can become inactive
after milestone-2. In that case, TC needs to discussion it
in details and considering all risk take decision. Modifying
the process doc to explicitly cover this case just in case
if we see project not fixing security bug or other critical
issue in project after m-2.

- https://etherpad.opendev.org/p/apr2024-ptg-os-tc#L303

Change-Id: I0c3908c93bc12db1fb8ef0d2c433352a83e49554