Add openstack-helm-plugin git repository
The repository is to contain the code
of the OSH plugin for Helm CLI tool.

Helm allows users install plugins from git
repos but it assumes cloning the whole git
repo to the user's machine which is not
desirable.

This repo will contain only a small portion
of the OSH code which provides some convenience
to users when they deploy Openstack using
OSH charts.

Depends-On: I1ec3e19c0dd8c83382e803525dd0b9612cfa75f8
Change-Id: Ib4b5e3baa22f87a19b3f75952eb8b587acd8283b