Move distributed-project-leadership model into doc
Distributed project leadership model details are written
in resolution only
- https://governance.openstack.org/tc/resolutions/20200803-distributed-project-leadership.html

but many times we need to gradually develop/amend this model for
example, in vPTG we discussed adding a monitoring policy for
DPL liaisons
- https://etherpad.opendev.org/p/apr2024-ptg-os-tc#L297

It is not easy to amend resolution for such changes and all the amendment
will be distributed in different resolutions which will make it hard to
have the latest policy/document at single place.

The content of the resolution is not changed in the new reference doc and
any further amendment will be done later.

Change-Id: I5518308b6d3f9e83547c43c30fe00b2f1972fdb5