Retire Murano project
Murano project is in inactive list since 2024.1 and
have an open critical bug.
- https://governance.openstack.org/tc/reference/emerging-technology-and-inactive-projects.html#current-inactive-projects

and there is no leader in 2024.2 cycle.
- https://etherpad.opendev.org/p/2024.2-leaderless

There is no response on last call for maintainers and retirement
notification:
- https://lists.openstack.org/archives/list/openstack-discuss@lists.openstack.org/thread/U2HITVJSZ4TKGTA7LEAODBIUZBCYGBZX/

Seeing project inactivities, open critical bug, and no maintainers
for this project, this commit propose to retire this project.

Change-Id: I6fc0070586b83c93398f3c2837e45974b23ba249