Retire Sahara project
Sahara project is in inactive list since 2024.1
- https://governance.openstack.org/tc/reference/emerging-technology-and-inactive-projects.html#current-inactive-projects

and there is no leader in 2024.2 cycle (TC appointed the PTL
for this project in 4 cycles out of last 5 cycles).
- https://etherpad.opendev.org/p/2024.2-leaderless

There is no response on last call for maintainers and retirement
notification:
- https://lists.openstack.org/archives/list/openstack-discuss@lists.openstack.org/thread/24M3WQXDE2DYN3GWDFHY52FNGNIXN4N3/

Seeing project inactivities, open critical bug, and no maintainers
for this project, this commit propose to retire this project.

Change-Id: Ibaeaf19a6ff8b8e07d56d7a6f51a9f95b8f1e28c