Clarify the project retirement vs project's repository retirement
In TC house rule, we introduced the project-updates topic so that
we can speed up the process for updates (including addition or retirement
of repository) in existing projects. But this was not meant for the
retirement of complete project. Any retirement of complete project
or a new project application are the big change in OpenStack umbrella
and should require the detailed discussion. That is why it should fall
under formal-vote topic.

For retiring the projct, we have been using the mixed topic where
sometime we use formal-vote and sometime project-updates.

Clarfiy the same in our house rule doc also.

Change-Id: I591c92fb84b5ca22905cf703be3c27f690899513