Add retired-on date for a few missing projects
When we retire the entire project team, we need to add the
'retired-on' flag and the date of retirement. A few projects
TripleO, infra, and Fuel project are retired completly but
do not have retired-on date for project team.

Also, adding repo info for refstack.

This will help fixing the validate-legacy script to detect the
project retirement correctly.
- https://review.opendev.org/c/openstack/governance/+/920848

Change-Id: I966e5ee1cb5c070757d62b1b2ab7ba38028e8caa