Update criteria for the inactive projects to become active again
This was discussed during the TC session in the vPTG in April 2024 that
we should raise the bar for an inactive projects to become again active
official OpenStack projects.
This patch updates the emerging technology and inactive projects
document to require from inactive projects that they need to meet all
the requirements to be active for at least one full cycle and only then
in next cycle such project can be considered by the TC to be active
again.

Change-Id: Ie16888099f1ed938b7753808984c7368ea2300a6