Fixing the validate-legacy script
validate-legacy script supposed to verify if repository is
retired properly or not basically keep only two file .gitreview
and README.rst[1].

Currently this script only checks the partially retired project
and not if project team itself is retired which end up having a
incorrect retirement. Example, recently retired project senlin:

  Even the repo content removal was not correct
  - https://review.opendev.org/c/openstack/solum/+/920201

  the below governance change still passed and this script did not complain
  - https://review.opendev.org/c/openstack/governance/+/919347/

Fixing it so that it cover both case partial as well as complete retirement.
[1] https://docs.openstack.org/project-team-guide/repository.html#step-2-remove-project-content

Ref: https://etherpad.opendev.org/p/tc-retirement-cleanup

Change-Id: I1d5048ec38fa1af625a32e0f0885e9a3c3f46d0e