Adding a rule to manage the affiliation diversity requirement in the TC
Last year, we moved the TC affiliation diversity requirement from
the Board bylaw to the TC charter, but we did not define the rules
to handle a violation of the diversity requirement.

During the 2024.2 TC election, we encountered a situation where the
diversity requirement was violated[1]. Due to the lack of defined
rules in the TC charter, we were unable to effectively resolve the
situation, resulting in an undesirable outcome.

As discussed at the PTG[2], we need to define some rules in the
charter to better handle such situations in the future. To keep it
simple, the proposal here is to select another affiliation member
(who does not violate the affiliation diversity) based on their
ranking in the election voting. This rule is align with the Board
diversity requirement voilation rules[3].

[1] https://governance.openstack.org/tc/resolutions/20240322-adjudication-of-2024.2-tc-election.html
[2] https://etherpad.opendev.org/p/apr2024-ptg-os-tc#L219
[3] https://openinfra.dev/legal/bylaws (Section 4.15 Director Diversity)

Change-Id: I1857cd95991abaad7fcdd7657d88eff7f44631ab