Inactive state extensions: Freezer
Sometimes, the project need more time to become active, and TC
can discuss their extension to stay Inactive. Adding the project
Inactive state extension in the timeline section (basically
"entered" to "entered or extended"). It means extension cycle
will be considered as the latest cycle for the Inactive project
timeline/next step.

Freezer project became inactive in 2024.1 cycle, and the
new maintainers took over the project maintenance. They
are working on making project active but need more time.
This proposes the extension for this project to stay inactive
for this cycle.

Change-Id: I7a5b3cab1323a20225b929e2002dabec3eebdce6