Add ansible-role-httpd repo to OSA-owned projects
In order to unify approach on handling Apache Web server deployment
and configuration team come up with idea to produce a separate role
for that as we have failed with finding existing one which would meet
our intended "pluggable" design.

Based on that a new repo/project is needed as we intend to make
resulting role not opinionated and re-usable outside of OSA scope.

Depends-On: https://review.opendev.org/c/openstack/project-config/+/935693
Change-Id: I2da92910e2897da10283bb93fcb6309c70d09d6e