Allow more than 2 weeks for elections
In the 2024.2 election cycle, the polling period
was three weeks. We noticed a small uptick in
the voting numbers:

- 2025.1: (21.4% voter) Voting period was three weeks
284 authorized voters, votes cast: 61
- 2024.2 (19.1% voter) Voting period was two weeks
298 authorized voters, votes cast: 57
- 2024.1: (16.5% voter) Voting period was two weeks
309 authorized voters, Actual votes cast: 51

While the reason for this uptick cannot solely be
attributed to the change in window, the TC expressed
their preference that the extended voting period
did not hurt. It is allowed for this voting
period to overlap with the week set aside for
TC campaigning - any early voters have the ability
to change their minds and their votes before
polling ends.

Change-Id: I45e12ce715708eb67449dc73db8cffcf3cc70234