Add separate freezer-api deliverable
freezer-api was considered as same deliverable as freezer (and
freezer-dr) in the past, to release with the same version, etc.
(details in this patch: Ib63712683c96587e6a7498d28cbd68edb6ddc367)

Freezer was retired and now it is re-activated and was added as a
separate deliverable in releases repository. To be able to release
freezer-api we have to move freezer-api as a separate deliverable
in projects.yaml as well.

Change-Id: I1fe857c796a89c539055a0a801e07891af772072