Deprecate openstack-ansible-tests repo
Tests repo was used for functional tests of OSA which project
mainly moved from around Train-Ussuri release. Some repositories
were still leveraging the repository for various thing up to Epoxy
release.

At the moment repository has no value, has broken CI and can
hardly be used with it's original concept as it was proved as not
efficient enough and causing constant circular dependencies,
while depends-on was not properly respected.

Change-Id: I4749eff103ee97344b7b42a9f73a045782be7a57