Clarify actions when no elections are required
During the nomination window for elections of project
team leads and technical committee members, if there aren't
enough nominees to warrant elections, we must declare election
results and forego the voting window that was set aside.

Doing this would mean that the term of the new TC could
begin a bit sooner (i.e., by up to 3 weeks sooner).
However, PTLs are elected in advance of a release cycle.

Change-Id: I26836bc05c9ec40bdc01e1134ad76f4dcce00bd9