Retire Monasca project
The Monasca project has been inactive and the team has decided
to retire the project. This change moves the project from
projects.yaml to legacy.yaml to formally retire it.

This is Step 3 of the repository retirement process as outlined
in the OpenStack Project Team Guide.

All Monasca repositories will be retired:
- monasca-api
- monasca-log-api (already deprecated in Ussuri)
- monasca-events-api
- monasca-specs
- monasca-notification
- monasca-persister
- monasca-tempest-plugin
- monasca-thresh
- monasca-common
- monasca-ui
- python-monascaclient
- monasca-agent
- monasca-statsd
- monasca-ceilometer (already deprecated in Ussuri)
- monasca-transform (already deprecated in Victoria)
- monasca-grafana-datasource
- monasca-kibana-plugin

Depends-On: https://review.opendev.org/c/openstack/project-config/+/957063
Change-Id: I3cb522ce8f51424b64e93c1efaf0dfd1781cd5ac
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>