Define testing runtime for 2026.1 release
This is a SLURP release, and upgrade compatibility
will require us to maintain testing with verions
of Operating Systems supported in the 2025.1 release.

Add python 3.14 testing with non-voting jobs, and
continue support for python3.10. It will go EOL
six months after we ship the 2026.1 release; however
our PTI guideline requires us not to end support
for actively maintained versions of python.

Change-Id: Ifa6f5b91472fe18e8ef2dc929238872484c75d2d
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>