Use DPL model for  watcher project leadership for 2026.1 release
Marios is no longer working on OpenStack project, so we are removing
him. Chandan, nd Joan had agreed to help with release. Douglas will
be helping on tact-sig side.

Sean will be helping on release side.

TODO: We need one tc liasion since  Ghanshyam Maan is not
running for this release.

Change-Id: Id161531af8f1dc083a321eac359f7178fb108abf
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>