Show inactive project status in project.yaml
Currently, if any project goes inactive, we mark
it as inactive in the below doc
- https://governance.openstack.org/tc/reference/emerging-technology-and-inactive-projects.html#current-inactive-projects

which is not very visible to everyone and official list
of projects (projects.yaml) does not provide the project
status info. I am adding the new field 'status' in
projects.yaml which can be added for the inactive project
with 'inactive' value. Unless any project is marked as
'inactive', everything else in projects.yaml can be considered
as active.

Also, updating the existing inactive project status.

Change-Id: Iee27669bb3a03befa06dc287be394063c0df5818
Signed-off-by: Ghanshyam Maan <gmaan@ghanshyammann.com>