Retire metalsmith project from bare metal program
On October 31, 2023, the Ironic team documented and notified
metalsmith users that the project was going into maintenance
mode ( e4fd02fa30164de00bc5a354af954b503f42c89b ).

Due to broken CI and a general lack of interest from maintainers
and users, we are retiring the metalsmith project.

This was communicated to stakeholders most recently with this
mailing list post:
https://lists.openstack.org/archives/list/openstack-discuss@lists.openstack.org/message/WFCMET4UJDW56YJKWGA5DSIB37LZUR4B/

Depends-On: https://review.opendev.org/c/openstack/metalsmith/+/963427
Signed-off-by: Jay Faulkner <jay@jvf.cc>
Change-Id: I0463b304ebad8e76e825f1d6d4afd522fbf2370c