Mark Vitrage project-team inactive
This team was deemed leaderless in the 2026.1
elections; and we've not managed to find any
PTL volunteers. We can assess the status of
the deliverables after M-2 of the 2026.1 release,
however, the release team is advised not to
consider any procedural releases for vitrage
deliverables for 2026.1 in this inactive state.

Change-Id: I263fc5c8259152ef58f02652181e08f14c2f56e5
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>