Mark Venus project-team inactive
This team was deemed leaderless in the 2026.1
elections; and we've not managed to find any
PTL volunteers. We can assess the status of
the deliverables after M-2 of the 2026.1 release,
however, the release team is advised not to
consider any procedural releases for vitrage
deliverables for 2026.1 in this inactive state.

Change-Id: I48e94e51728220138879965df5e5f6056bed488a
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>