Fix SVG badge generation
SVG badges were incorrectly generated since the move to getbbox() [1].
Unlike getsize(), getbbox() returns a bounding box as four floats [2]:
(left, top, right, bottom).

The `left` coordinate was always 0, which resulted in the width being
set to PADDING (8 pixels), making the text unreadable.

We need to use the `right` coordinate instead. For extra safety in case
`left` is ever not zero, we measure the distance between the `left` and
`right` coordinates.

[1] https://review.opendev.org/c/openstack/governance/+/935967
[2] https://pillow.readthedocs.io/en/stable/reference/ImageFont.html#PIL.ImageFont.FreeTypeFont.getbbox

Change-Id: I57d7720f441dca0f41dbf7d5e11216a03b3b585f
Signed-off-by: Pierre Riteau <pierre@stackhpc.com>