Move keystoneauth under OpenStackSDK governance
ksa is a core dependency of SDK (and by extension, OSC), and is
typically used in tandem with it or legacy clients. It is expected that
these libraries will continue to develop in tandem, with a different
level of investment (and potentially release cadence) to keystone
itself. OpenStackSDK feels like a more natural home for this nowadays.

Change-Id: Id1a5770863f8dd5982c8d92aa190a02820f4b19e
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>